# lib/api-zod/src/generated/types/bannerLang.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Hero text for one language
 */
export interface BannerLang {
  hero_badge?: string;
  hero_h1a?: string;
  hero_h1b?: string;
  hero_sub?: string;
  hero_cta1?: string;
  hero_cta2?: string;
}