# lib/api-zod/src/generated/types/bannerSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BannerLang } from './bannerLang';

export interface BannerSettings {
  en?: BannerLang;
  ar?: BannerLang;
}