# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BannerSettings } from './bannerSettings';
import type { ThemeSettings } from './themeSettings';

export interface SiteSettings {
  banners: BannerSettings;
  theme: ThemeSettings;
}