# lib/api-zod/src/generated/types/themeSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * CSS colour overrides (hex values)
 */
export interface ThemeSettings {
  primary?: string;
  secondary?: string;
  accent?: string;
}